smart_charging, types: Move ReportedChargingProfiles
Move the ReportedChargingProfiles struct to types.hpp where it can
be used in e.g. DatabaseHandler without causing recursive imports.
